refactor: :recycle: move MeiliSearch code to `meilisearch` package and only register commands/hooks if MeiliSearch environment variables are set
